Move the corrector loss objects beside the losses they compose
CorrectorLoss, StepOutputLoss, and StepOutputLossOutput move to fme/core/loss.py;
fme/core/corrector/loss.py becomes loss_config.py, holding only the configs. The
dependency runs one way: the config module imports CorrectorLoss, and
fme/core/loss.py still imports nothing from fme/core/corrector/. Tests follow
their modules — loss behavior in fme/core/test_loss.py against directly
constructed objects, config and build validation in test_loss_config.py.

"Corrector regularization" now names only the feature; the tensors it produces
are penalties. So penalty_weight, CorrectorLoss.penalty, corrector_penalty, and
get_corrector_penalty_losses, in code and metric keys alike.

The penalty metrics move under a loss/ prefix, and the train aggregator's filter
widens to admit it. Both aggregators select which of TrainOutput.metrics reach
PerStepLossAggregator by key prefix, and a bare corrector_regularization key
passed neither: TrainOutput.metrics held it and nothing logged it. The keys are
now loss/corrector_penalty and loss/corrector_penalty_step_{i}, logged as
train/mean/loss/..., with per-channel penalties under corrector_penalty/<var>
landing at train/mean/loss/corrector_penalty/<var>. A test asserts on get_logs
rather than on TrainOutput.metrics, since only the former catches a dropped key.

_accumulate_step_loss returns a _StepLossOutput instead of mutating metrics,
weighted_sums, and total_counts in place, and _accumulate_loss does the
accumulating — it already owns the step loop and the containers. Threading the
corrector deltas through the old shape would have taken that signature to eleven
parameters, four of them out-params; returning instead leaves it four parameters
smaller than on main, and every metric key this PR touches now has one write
site. _accumulate_loss returns an _AccumulateLossOutput splitting the per-step
predictions from the detached reporting half, whose corrector_penalty property
means over the steps that produced one.

Private classes and methods carry no docstrings: a short inline comment on a
field or at a seam is the documentation. The reasoning a reader needs moved to
the public StepOutputLoss docstring, and the seven-line hard-boundary comment in
the step loop is one line.
