+++ fme/ace/aggregator/test_train.py
@@ -334,3 +334,54 @@ def test_aggregator_ensemble_metrics_skipped_with_single_member():
     assert "train/ensemble/crps/a" not in logs
     assert "train/ensemble/ssr_bias/a" not in logs
     assert "train/ensemble/ensemble_mean_rmse/a" not in logs
+
+
+def test_corrector_penalty_metrics_reach_the_logs():
+    # GOAL: a TrainOutput carrying loss/corrector_penalty,
+    # loss/corrector_penalty_step_0, and a corrector_penalty/<var> per-channel
+    # entry logs all three under train/mean/loss/. Asserting on
+    # TrainOutput.metrics alone cannot catch the dropped-by-filter case, so this
+    # asserts on get_logs output.
+    batch_size = 4
+    n_ensemble = 1
+    n_time = 2
+    nx, ny = 2, 2
+    device = get_device()
+    gridded_operations = LatLonOperations(
+        area_weights=torch.ones(nx, ny, device=device)
+    )
+    config = TrainAggregatorConfig(
+        spherical_power_spectrum=False, weighted_rmse=False, per_channel_loss=True
+    )
+    agg = TrainAggregator(config=config, operations=gridded_operations)
+    agg.record_batch(
+        batch=TrainOutput(
+            metrics={
+                "loss": torch.tensor(1.0, device=device),
+                "loss/corrector_penalty": torch.tensor(0.25, device=device),
+                "loss/corrector_penalty_step_0": torch.tensor(0.5, device=device),
+            },
+            target_data=EnsembleTensorDict(
+                {"a": torch.randn(batch_size, 1, n_time, nx, ny, device=device)},
+            ),
+            gen_data=EnsembleTensorDict(
+                {
+                    "a": torch.randn(
+                        batch_size, n_ensemble, n_time, nx, ny, device=device
+                    )
+                },
+            ),
+            time=xr.DataArray(np.zeros((batch_size, n_time)), dims=["sample", "time"]),
+            normalize=lambda x: x,
+            per_channel_losses={
+                "a": ChannelLossInfo(loss=torch.tensor(0.5, device=device), count=4),
+                "corrector_penalty/a": ChannelLossInfo(
+                    loss=torch.tensor(0.125, device=device), count=4
+                ),
+            },
+        ),
+    )
+    logs = agg.get_logs(label="train")
+    assert logs["train/mean/loss/corrector_penalty"] == 0.25
+    assert logs["train/mean/loss/corrector_penalty_step_0"] == 0.5
+    assert logs["train/mean/loss/corrector_penalty/a"] == 0.125
