+++ fme/core/test_loss.py
@@ -6,6 +6,7 @@ from fme.core.device import get_device
 from fme.core.gridded_ops import GriddedOperations, LatLonOperations
 from fme.core.loss import (
     AreaWeightedMSELoss,
+    CorrectorLoss,
     CRPSLoss,
     EnergyScoreLoss,
     EnsembleComponentLoss,
@@ -18,7 +19,9 @@ from fme.core.loss import (
     SpectralWhitening,
     SpectralWhiteningConfig,
     StandardLoss,
+    StepLoss,
     StepLossConfig,
+    StepOutputLoss,
     VariableWeightingLoss,
     WeightedMappingLoss,
     _construct_weight_tensor,
@@ -997,3 +1000,217 @@ def test_ensemble_loss_with_whitening_builds_and_runs(whitening):
     out = loss(x, y)
     assert all(isinstance(c, LossComponent) for c in out)
     assert torch.isfinite(_components_total(out))
+
+
+# Corrector-delta loss objects. Deterministic setup: fixed prediction/target/delta
+# dicts, a normalizer with known means/stds and an MSE step loss, so every penalty
+# below is computable by hand.
+_CORRECTOR_SHAPE = (2, 2, 2)  # (batch, lat, lon)
+_CORRECTOR_NAMES = ["a", "b_0", "b_1"]
+_CORRECTOR_MEANS = {"a": 1.0, "b_0": -2.0, "b_1": 0.5}
+_CORRECTOR_STDS = {"a": 2.0, "b_0": 0.5, "b_1": 4.0}
+
+
+def _fixed(offset: float, scale: float) -> torch.Tensor:
+    n = _CORRECTOR_SHAPE[0] * _CORRECTOR_SHAPE[1] * _CORRECTOR_SHAPE[2]
+    values = offset + scale * torch.arange(n, dtype=torch.float32)
+    return values.reshape(_CORRECTOR_SHAPE).to(get_device())
+
+
+def _predict_dict() -> dict[str, torch.Tensor]:
+    return {
+        "a": _fixed(0.0, 0.25),
+        "b_0": _fixed(1.0, -0.5),
+        "b_1": _fixed(-3.0, 0.75),
+    }
+
+
+def _target_dict() -> dict[str, torch.Tensor]:
+    return {
+        "a": _fixed(0.5, 0.2),
+        "b_0": _fixed(-1.0, 0.4),
+        "b_1": _fixed(2.0, -0.3),
+    }
+
+
+def _delta_dict() -> dict[str, torch.Tensor]:
+    return {
+        "a": _fixed(0.1, 0.05),
+        "b_0": _fixed(-0.2, 0.1),
+        "b_1": _fixed(0.3, -0.15),
+    }
+
+
+def _corrector_normalizer() -> StandardNormalizer:
+    return StandardNormalizer(
+        means={
+            name: torch.as_tensor(_CORRECTOR_MEANS[name]) for name in _CORRECTOR_NAMES
+        },
+        stds={
+            name: torch.as_tensor(_CORRECTOR_STDS[name]) for name in _CORRECTOR_NAMES
+        },
+    )
+
+
+def _corrector_step_loss() -> StepLoss:
+    return StepLossConfig(type="MSE").build(
+        gridded_ops=None,
+        out_names=list(_CORRECTOR_NAMES),
+        normalizer=_corrector_normalizer(),
+    )
+
+
+def _corrector_loss(
+    precorrector_names: list[str] | None = None,
+    regularizer_names: list[str] | None = None,
+    penalty_weight: float = 1.0,
+) -> CorrectorLoss:
+    regularizer = None
+    if regularizer_names is not None:
+        regularizer = WeightedMappingLoss(
+            loss=LossConfig(type="MSE").build(gridded_operations=None),
+            weights={},
+            out_names=regularizer_names,
+            normalizer=_corrector_normalizer(),
+            channel_dim=-3,
+        )
+    return CorrectorLoss(
+        precorrector_names=precorrector_names,
+        regularizer=regularizer,
+        penalty_weight=penalty_weight,
+    )
+
+
+def _expected_penalty(deltas, names) -> torch.Tensor:
+    per_channel = []
+    for name in sorted(names):
+        normalized = torch.nan_to_num(deltas[name], nan=0.0) / _CORRECTOR_STDS[name]
+        per_channel.append((normalized**2).mean())
+    return torch.stack(per_channel).mean()
+
+
+def test_step_output_loss_without_corrector_loss_unchanged():
+    # GOAL: with corrector_loss None, total() and get_channel_losses() match a
+    # bare StepLoss, deltas ignored.
+    predict, target, deltas = _predict_dict(), _target_dict(), _delta_dict()
+    bare = _corrector_step_loss()(predict, target, 0)
+    result = StepOutputLoss(_corrector_step_loss(), None)(
+        predict, target, 0, deltas=deltas
+    )
+    assert result.corrector_penalty is None
+    torch.testing.assert_close(result.total(), bare.total())
+    assert result.get_corrector_penalty_losses() == {}
+    expected = bare.get_channel_losses()
+    actual = result.get_channel_losses()
+    assert set(actual) == set(expected)
+    for name in expected:
+        torch.testing.assert_close(actual[name].loss, expected[name].loss)
+
+
+def test_pre_corrector_outputs_selected_only():
+    # GOAL: the main loss sees output − delta for the configured names only;
+    # other keys use the network output as-is; targets untouched.
+    predict, target, deltas = _predict_dict(), _target_dict(), _delta_dict()
+    corrector_loss = _corrector_loss(precorrector_names=["a"])
+    net_output = corrector_loss.pre_corrector_outputs(predict, deltas)
+    torch.testing.assert_close(net_output["a"], predict["a"] - deltas["a"])
+    for name in ("b_0", "b_1"):
+        torch.testing.assert_close(net_output[name], predict[name])
+    result = StepOutputLoss(_corrector_step_loss(), corrector_loss)(
+        predict, target, 0, deltas=deltas
+    )
+    expected = _corrector_step_loss()(net_output, target, 0)
+    torch.testing.assert_close(result.main.total(), expected.total())
+    assert result.corrector_penalty is None
+    # the target dict is never modified
+    for name, value in _target_dict().items():
+        torch.testing.assert_close(target[name], value)
+
+
+def test_penalty_analytic_value():
+    # GOAL: penalty equals the hand-computed mean of (delta/std)^2 — normalizer
+    # means cancel against the zeros target.
+    deltas = _delta_dict()
+    penalty = _corrector_loss(regularizer_names=_CORRECTOR_NAMES).penalty(deltas)
+    assert penalty is not None
+    torch.testing.assert_close(
+        penalty.total(), _expected_penalty(deltas, _CORRECTOR_NAMES)
+    )
+
+
+def test_penalty_masked_points_drop():
+    # GOAL: NaN-filled delta points contribute nothing; penalty and gradients
+    # stay finite.
+    deltas = _delta_dict()
+    masked = deltas["a"].clone()
+    masked[0] = torch.nan
+    masked.requires_grad_(True)
+    deltas["a"] = masked
+    penalty = _corrector_loss(regularizer_names=_CORRECTOR_NAMES).penalty(deltas)
+    assert penalty is not None
+    total = penalty.total()
+    assert torch.isfinite(total)
+    torch.testing.assert_close(total, _expected_penalty(deltas, _CORRECTOR_NAMES))
+    total.backward()
+    assert masked.grad is not None
+    assert torch.isfinite(masked.grad).all()
+    assert (masked.grad[0] == 0.0).all()
+
+
+def test_total_and_channel_decomposition():
+    # GOAL: total() == main.total() + weight * penalty.total();
+    # get_channel_losses() is main-only; get_corrector_penalty_losses() covers
+    # exactly the selected names.
+    predict, target, deltas = _predict_dict(), _target_dict(), _delta_dict()
+    corrector_loss = _corrector_loss(
+        precorrector_names=["a"],
+        regularizer_names=["b_0", "b_1"],
+        penalty_weight=3.0,
+    )
+    result = StepOutputLoss(_corrector_step_loss(), corrector_loss)(
+        predict, target, 0, deltas=deltas
+    )
+    assert result.corrector_penalty is not None
+    torch.testing.assert_close(
+        result.total(),
+        result.main.total() + 3.0 * result.corrector_penalty.total(),
+    )
+    assert list(result.get_channel_losses()) == _CORRECTOR_NAMES
+    penalty_channels = result.get_corrector_penalty_losses()
+    assert list(penalty_channels) == ["b_0", "b_1"]
+    # per-channel penalties are unweighted
+    torch.testing.assert_close(
+        torch.stack([info.loss for info in penalty_channels.values()]).mean(),
+        _expected_penalty(deltas, ["b_0", "b_1"]),
+    )
+
+
+@pytest.mark.parametrize("feature", ["precorrector_optimization", "regularization"])
+def test_missing_selected_delta_raises(feature):
+    # GOAL: a non-empty delta dict lacking a selected name raises.
+    predict, target = _predict_dict(), _target_dict()
+    deltas = {k: v for k, v in _delta_dict().items() if k != "a"}
+    if feature == "precorrector_optimization":
+        corrector_loss = _corrector_loss(precorrector_names=["a"])
+    else:
+        corrector_loss = _corrector_loss(regularizer_names=["a"])
+    loss = StepOutputLoss(_corrector_step_loss(), corrector_loss)
+    with pytest.raises(ValueError, match="produced no delta"):
+        loss(predict, target, 0, deltas=deltas)
+
+
+@pytest.mark.parametrize("deltas", [None, {}])
+def test_empty_deltas_inert(deltas):
+    # GOAL: empty deltas ⇒ no pre-corrector swap, no penalty; result matches
+    # the unconfigured case.
+    predict, target = _predict_dict(), _target_dict()
+    corrector_loss = _corrector_loss(
+        precorrector_names=["a"], regularizer_names=["b_0", "b_1"]
+    )
+    result = StepOutputLoss(_corrector_step_loss(), corrector_loss)(
+        predict, target, 0, deltas=deltas
+    )
+    unconfigured = StepOutputLoss(_corrector_step_loss(), None)(predict, target, 0)
+    assert result.corrector_penalty is None
+    assert result.get_corrector_penalty_losses() == {}
+    torch.testing.assert_close(result.total(), unconfigured.total())
